services/ml-detection/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiokafka==0.10.0
numpy==1.26.3
torch>=2.2.0
onnxruntime>=1.17.0
onnx>=1.15.0
psycopg2-binary==2.9.9
python-json-logger==2.0.7
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
